+++ NEWS.md
@@ -3,6 +3,10 @@
 ## Enhancements
 - [Sort the rows by `country_code` and `reporting_year` in the pip response.](https://github.com/PIP-Technical-Team/pipapi/issues/248)
 - [Fix casing making pip call case insensitive](https://github.com/PIP-Technical-Team/pipapi/issues/120)
+- Region codes can now be passed directly to the `country` query parameter to 
+return all countries pertaining to the specified region
+- the `/valid-params` endpoint gains an `endpoint` parameter that allows to only
+return parameters that are relevant to the specified endpoint
 
 
 # pipapi 1.0.0
@@ -26,6 +30,7 @@
 - Add country and region name to /pip response 
 - Make sure latest data version is available as a specific version (not just as "latest_relase")
 - Use `{fs}` functions rather than base R's. 
+- [Sort the rows by `country_code` and `reporting_year` in the pip response.](https://github.com/PIP-Technical-Team/pipapi/issues/248) 
 
 ## Bug fixes